kernelinstaller: Add zip input and split kernel+dtb pairs, plus fix fd leaks

+++ kernelinstaller/main.go
@@ -1,9 +1,9 @@
 package main
 
 import (
+	"archive/zip"
 	"fmt"
 	"io"
-	"io/ioutil"
 	"os"
 	"os/exec"
 	"path/filepath"
@@ -17,6 +17,8 @@ var (
 	wd, mb string
 	kernel, dtb string
 	boot, vendorboot string
+
+	BUFFERSIZE int64 = 4096
 )
 
 func init() {
@@ -61,13 +63,6 @@ func check(err error) {
 }
 
 func main() {
-	/*TODO:
-	- Detect if kernel is an archive (ZIP, RAR, 7Z), and if so, unpack it and search for a kernel and dtb pair
-	- Detect if kernel is an Android boot image, and if so, unpack the ramdisk and search for a kernel and dtb pair,
-		otherwise use the kernel and dtb from the boot image
-	- Detect if kernel is not a kernel and error out
-	*/
-	
 	log("Backing up boot to [/sdcard/boot.img]...")
 	check(cp(boot, "/sdcard/boot.img"))
 
@@ -94,25 +89,58 @@ func main() {
 		typeKernel = "linux"
 	} else if strings.Contains(typeKernel, "Android bootimg") {
 		typeKernel = "boot"
+	} else if strings.Contains(typeKernel, "Zip archive") {
+		typeKernel = "zip"
 	} else {
 		check(fmt.Errorf("[%s] is not an Android kernel: %s", kernel, typeKernel))
 	}
 	log("Successfully validated kernel image as " + file(kernel))
-	
+
 	if typeKernel == "linux" {
 		log("Nothing to do for kernel")
-	} else if typeKernel == "boot" {
-		log("Unpacking kernel image to [" + wd+"kernel]...")
-		check(os.MkdirAll(wd+"kernel/rd", 0644))
-		check(magiskboot(wd+"kernel", "unpack", "-h", kernel))
+	} else if typeKernel == "boot" || typeKernel == "zip" {
+		check(os.MkdirAll(wd+"kernel/tmp", 0644))
 
-		log("Unpacking kernel image ramdisk to [" + wd+"kernel/rd]...")
-		check(magiskboot(wd+"kernel/rd", "cpio", wd+"kernel/ramdisk.cpio", "extract"))
+		switch typeKernel {
+		case "boot":
+			log("Unpacking kernel image to [" + wd+"kernel]...")
+			check(magiskboot(wd+"kernel", "unpack", "-h", kernel))
+
+			log("Unpacking kernel image ramdisk to [" + wd+"kernel/tmp]...")
+			check(magiskboot(wd+"kernel/tmp", "cpio", wd+"kernel/ramdisk.cpio", "extract"))
+		case "zip":
+			log("Unpacking kernel zip to [" + wd+"kernel/tmp]...")
+			archive, err := zip.OpenReader(kernel)
+			check(err)
+			defer archive.Close()
+
+			for _, f := range archive.File {
+				filePath := filepath.Join(wd+"kernel/tmp", f.Name)
+
+				if f.FileInfo().IsDir() {
+					os.MkdirAll(filePath, os.ModePerm)
+					continue
+				}
+
+				check(os.MkdirAll(filepath.Dir(filePath), os.ModePerm))
+				dstFile, err := os.OpenFile(filePath, os.O_WRONLY|os.O_CREATE|os.O_TRUNC, f.Mode())
+				check(err)
+
+				fileInArchive, err := f.Open()
+				check(err)
+
+				_, err = io.Copy(dstFile, fileInArchive)
+				check(err)
+
+				dstFile.Close()
+				fileInArchive.Close()
+			}
+		}
 
 		log("Walking for kernel and dtb...")
 		kernel = ""
 		dtb = ""
-		err := filepath.Walk(wd+"kernel/rd", func(path string, info os.FileInfo, err error) error {
+		err := filepath.Walk(wd+"kernel/tmp", func(path string, info os.FileInfo, err error) error {
 			typeFile := file(path)
 			if strings.Contains(typeFile, "compressed data") {
 				check(magiskboot(wd+"kernel", "decompress", path, path+".decompressed"))
@@ -125,6 +153,10 @@ func main() {
 					check(fmt.Errorf("TODO: kernel choice not yet supported"))
 				}
 				kernel = path
+				if strings.Contains(path, "dtb") {
+					log("Found kernel+dtb: " + path)
+					return io.EOF
+				}
 				log("Found kernel: " + path)
 			}
 			if strings.Contains(typeFile, "Device Tree Blob") {
@@ -151,7 +183,7 @@ func main() {
 			kernel = wd+"kernel/kernel"
 			dtb = wd+"kernel/dtb"
 		}
-			
+
 		typeKernel = file(kernel)
 		if !strings.Contains(typeKernel, "Linux kernel") {
 			check(fmt.Errorf("[%s] is not an Android kernel: %s", kernel, typeKernel))
@@ -159,56 +191,92 @@ func main() {
 		log("Successfully validated kernel as " + typeKernel)
 
 		log("Cleaning up unpacked kernel image...")
-		check(cp(kernel, wd+"kernel.tmp"))
-		check(cp(dtb, wd+"dtb.tmp"))
+		check(os.Rename(kernel, wd+"kernel.tmp"))
 		kernel = wd+"kernel.tmp"
-		dtb = wd+"dtb.tmp"
+		if dtb != "" {
+			check(os.Rename(dtb, wd+"dtb.tmp"))
+			dtb = wd+"dtb.tmp"
+		}
 		check(os.RemoveAll(wd+"kernel"))
 	} else {
 		check(fmt.Errorf("What are we supposed to do, exactly?"))
 	}
 	
 	if dtb == "" {
-		check(fmt.Errorf("TODO: embedded dtbs and dtb scanning not yet supported"))
+		log("Trying to split kernel for dtb...")
+		check(os.MkdirAll(wd+"dtb", 0644))
+		if err := magiskboot(wd+"dtb", "split", kernel); err == nil {
+			kernel = wd+"dtb/kernel"
+			dtb = wd+"dtb/kernel_dtb"
+
+			typeKernel = file(kernel)
+			if !strings.Contains(typeKernel, "Linux kernel") {
+				check(fmt.Errorf("[%s] is not an Android kernel: %s", kernel, typeKernel))
+			}
+		}
 	}
-	typeDTB := file(dtb)
-	if !strings.Contains(typeDTB, "Device Tree Blob") {
-		check(fmt.Errorf("[%s] is not a Device Tree Blob: %s", dtb, typeDTB))
+	if dtb != "" {
+		typeDTB := file(dtb)
+		if !strings.Contains(typeDTB, "Device Tree Blob") {
+			check(fmt.Errorf("[%s] is not a Device Tree Blob: %s", dtb, typeDTB))
+		}
+		log("Successfully validated dtb as " + typeDTB)
+	} else {
+		log("No device tree blob found, ignoring...")
 	}
-	log("Successfully validated dtb as " + typeDTB)
 
 	log("Unpacking boot to [" + wd+"boot]...")
 	check(os.MkdirAll(wd+"boot", 0644))
 	check(magiskboot(wd+"boot", "unpack", "-n", boot))
 
 	log("Injecting kernel into boot...")
-	check(cp(kernel, wd+"boot/kernel"))
-	log("Injecting dtb into boot...")
-	check(cp(dtb, wd+"boot/dtb"))
+	check(os.Rename(kernel, wd+"boot/kernel"))
+	if dtb != "" {
+		log("Injecting dtb into boot...")
+		check(os.Rename(dtb, wd+"boot/dtb"))
+	} else if file(dtb) != "" {
+		log("Removing dtb from boot...")
+		check(os.RemoveAll(wd+"boot/dtb"))
+	}
 
 	log("Repacking boot...")
-	check(magiskboot(wd+"boot", "repack", "-n", boot, wd+"boot.img"))
+	check(magiskboot(wd+"boot", "repack", "-n", boot, wd+"new.b.img"))
 	log("Cleaning up unpacked boot...")
+	if dtb != "" {
+		check(os.Rename(wd+"boot/dtb", dtb))
+	}
 	check(os.RemoveAll(wd+"boot"))
 
-	log("Flashing boot...")
-	check(cp(wd+"boot.img", boot))
+	typeBoot = file(wd+"new.b.img")
+	if !strings.Contains(typeBoot, "Android bootimg") {
+		check(fmt.Errorf("Failed to repack boot"))
+	}
+	log("Successfully repacked boot as " + typeBoot)
 
-	if vendorboot != "" {
+	log("Flashing boot...")
+	check(cp(wd+"new.b.img", boot))
+
+	if vendorboot != "" && dtb != "" {
 		log("Unpacking vendor boot to [" + wd+"vendor_boot]...")
 		check(os.MkdirAll(wd+"vendor_boot", 0644))
 		check(magiskboot(wd+"vendor_boot", "unpack", "-n", vendorboot))
 
 		log("Injecting dtb into vendor boot...")
-		check(cp(dtb, wd+"vendor_boot/dtb"))
+		check(os.Rename(dtb, wd+"vendor_boot/dtb"))
 
 		log("Repacking vendor boot...")
-		check(magiskboot(wd+"vendor_boot", "repack", "-n", vendorboot, wd+"vendor_boot.img"))
+		check(magiskboot(wd+"vendor_boot", "repack", "-n", vendorboot, wd+"new.vb.img"))
 		log("Cleaning up unpacked vendor boot...")
 		check(os.RemoveAll(wd+"vendor_boot"))
 
+		typeVendorBoot := file(wd+"new.vb.img")
+		if !strings.Contains(typeVendorBoot, "data") {
+			check(fmt.Errorf("Failed to repack vendor boot"))
+		}
+		log("Successfully repacked vendor boot as " + typeVendorBoot)
+
 		log("Flashing vendor boot...")
-		check(cp(wd+"vendor_boot.img", vendorboot))
+		check(cp(wd+"new.vb.img", vendorboot))
 	}
 }
 
@@ -228,21 +296,43 @@ func run(prog, dir string, args ...string) error {
 }
 
 func cp(src, dst string) error {
-	file, err := os.OpenFile(src, os.O_RDONLY, os.ModeAppend)
+	source, err := os.Open(src)
 	if err != nil {
 		return err
 	}
-	data, err := ioutil.ReadAll(file)
+	defer source.Close()
+
+	destination, err := os.Create(dst)
 	if err != nil {
 		return err
 	}
-	return os.WriteFile(dst, data, 0644)
+	defer destination.Close()
+
+	if err != nil {
+		panic(err)
+	}
+
+	buf := make([]byte, BUFFERSIZE)
+	for {
+		n, err := source.Read(buf)
+		if err != nil && err != io.EOF {
+			return err
+		}
+		if n == 0 {
+			break
+		}
+
+		if _, err := destination.Write(buf[:n]); err != nil {
+			return err
+		}
+	}
+	return err
 }
 
 func file(path string) string {
 	fileType, err := exec.Command(wd+"/bin/file-" + runtime.GOARCH, "--magic-file", wd+"bin/magic.mgc", path).Output()
 	if err != nil {
-		return fmt.Sprintf("file: %v", err)
+		return ""
 	}
 
 	return string(fileType[len(path)+2:len(fileType)-1])
